Use Interface for Scheduling Selection
