+++ app/src/main/java/com/uni/stuttgart/ipvs/androidgateway/gateway/IScheduler.java
@@ -1,0 +1,6 @@
+package com.uni.stuttgart.ipvs.androidgateway.gateway;
+
+public interface IScheduler {
+    void start();
+    void stop();
+}
