+++ app/src/main/java/com/uni/stuttgart/ipvs/androidgateway/gateway/scheduler/ExhaustivePolling.java
@@ -1,40 +1,39 @@
-package com.uni.stuttgart.ipvs.androidgateway.gateway.scheduling;
+package com.uni.stuttgart.ipvs.androidgateway.gateway.scheduler;
 
 import android.bluetooth.BluetoothDevice;
 import android.content.Context;
 import android.content.Intent;
 import android.os.RemoteException;
+import android.util.Log;
 
 import com.uni.stuttgart.ipvs.androidgateway.gateway.GatewayService;
 import com.uni.stuttgart.ipvs.androidgateway.gateway.IGatewayService;
-import com.uni.stuttgart.ipvs.androidgateway.helper.PowerEstimator;
-import com.uni.stuttgart.ipvs.androidgateway.thread.EExecutionType;
+import com.uni.stuttgart.ipvs.androidgateway.gateway.IScheduler;
 import com.uni.stuttgart.ipvs.androidgateway.thread.ExecutionTask;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.Future;
 
-// implementation of Semaphore Scheduling Gateway Controller
-
-public class Semaphore {
-    private int SCAN_TIME; // set scanning and reading time to 10 seoonds
-    private int SCAN_TIME_HALF;
-    private int PROCESSING_TIME; // set processing time to 60 seconds
+// implementation of Scheduling using Exhaustive Polling
+public class ExhaustivePolling implements IScheduler {
 
     private IGatewayService iGatewayService;
-    private PowerEstimator mServicePE;
 
-    private boolean mBoundPE;
+    private int SCAN_TIME; // set scanning and reading time to 10 seoonds
+    private int SCAN_TIME_OTHER; // set scanning and reading time half of original scan time
+    private int PROCESSING_TIME; // set processing time to 60 seconds
+
     private boolean mScanning;
     private Context context;
     private boolean mProcessing;
+
     private int cycleCounter = 0;
-
     private ExecutionTask<Void> executionTask;
-    private Future<Void> future;
+    private Future<?> futureEP;
 
-    public Semaphore(Context context, boolean mProcessing, IGatewayService iGatewayService, ExecutionTask<Void> executionTask) {
+
+    public ExhaustivePolling(Context context, boolean mProcessing, IGatewayService iGatewayService, ExecutionTask<Void> executionTask) {
         this.context = context;
         this.mProcessing = mProcessing;
         this.iGatewayService = iGatewayService;
@@ -42,72 +41,89 @@ public class Semaphore {
 
         try {
             this.SCAN_TIME = iGatewayService.getTimeSettings("ScanningTime");
-            this.SCAN_TIME_HALF = iGatewayService.getTimeSettings("ScanningTime2");
+            this.SCAN_TIME_OTHER = iGatewayService.getTimeSettings("ScanningTime2");
             this.PROCESSING_TIME = iGatewayService.getTimeSettings("ProcessingTime");
         } catch (RemoteException e) {
             e.printStackTrace();
         }
     }
-    
+
     public void stop() {
         mProcessing = false;
-        future.cancel(true);
+        futureEP.cancel(true);
     }
 
     public void start() {
-        executionTask.setExecutionType(EExecutionType.SINGLE_THREAD_POOL);
-        future = executionTask.submitRunnable(new SemaphoreSchedulling());
+        mProcessing = true;
+        futureEP = executionTask.submitRunnable(new RunEPScheduling());
     }
 
-    private class SemaphoreSchedulling implements Runnable {
-
+    private class RunEPScheduling implements Runnable {
         @Override
         public void run() {
             while (!Thread.currentThread().isInterrupted() && mProcessing) {
+                // do Exhaustive Polling Part
                 cycleCounter++;
-                if(cycleCounter > 1) {broadcastClrScrn();}
+                if (cycleCounter > 1) {
+                    broadcastClrScrn();
+                }
                 broadcastUpdate("\n");
-                broadcastUpdate("Start new cycle...");
+                broadcastUpdate("Start new cycle");
                 broadcastUpdate("Cycle number " + cycleCounter);
 
                 try {
                     iGatewayService.setCycleCounter(cycleCounter);
+                    boolean isDataExist = iGatewayService.checkDevice(null);
+                    if (isDataExist) {
+                        // Devices are listed in DB
+                        List<String> devices = iGatewayService.getListActiveDevices();
+                        // search for known device listed in database
+                        for (String device : devices) {
+                            iGatewayService.startScanKnownDevices(device);
+                        }
 
-                    iGatewayService.startScan(SCAN_TIME);
-                    iGatewayService.stopScanning();
+                        // do normal scanning only for half of normal scanning time
+                        iGatewayService.startScan(SCAN_TIME_OTHER);
+                        iGatewayService.stopScan();
+                        mScanning = iGatewayService.getScanState();
+                    } else {
+                        // do normal scanning
 
-                    mScanning = iGatewayService.getScanState();
-
-                    if (!mProcessing) {
-                        future.cancel(true);
-                        Thread.currentThread().interrupt();
-                        return;
+                        iGatewayService.startScan(SCAN_TIME);
+                        iGatewayService.stopScan();
+                        mScanning = iGatewayService.getScanState();
                     }
 
                     List<BluetoothDevice> scanResults = iGatewayService.getScanResults();
                     iGatewayService.setScanResultNonVolatile(scanResults);
 
-                    // do Semaphore for Connecting method
+                    if (!mProcessing) {
+                        futureEP.cancel(true);
+                        Thread.currentThread().interrupt();
+                        return;
+                    }
+
+                    // do Connecting by using Semaphore
                     for (BluetoothDevice device : new ArrayList<BluetoothDevice>(scanResults)) {
-                        // only known manufacturer that will be used to connect
                         iGatewayService.broadcastServiceInterface("Start service interface");
                         iGatewayService.doConnect(device.getAddress());
 
                         if (!mProcessing) {
-                            future.cancel(true);
+                            futureEP.cancel(true);
                             Thread.currentThread().interrupt();
                             return;
                         }
-
                     }
 
                     if (!mProcessing) {
-                        future.cancel(true);
+                        futureEP.cancel(true);
                         Thread.currentThread().interrupt();
                         return;
                     }
-                } catch (RemoteException e) {
+                } catch (Exception e) {
                     e.printStackTrace();
+                } finally {
+                    Log.d("Thread", "Thread EP " + Thread.currentThread().getId() + " is interrupted");
                 }
             }
         }
