+++ app/src/main/java/com/uni/stuttgart/ipvs/androidgateway/gateway/scheduler/PriorityBasedWithWSM.java
@@ -1,69 +1,83 @@
-package com.uni.stuttgart.ipvs.androidgateway.gateway.scheduling;
+package com.uni.stuttgart.ipvs.androidgateway.gateway.scheduler;
 
 import android.bluetooth.BluetoothDevice;
 import android.content.Context;
 import android.content.Intent;
 import android.os.RemoteException;
+import android.util.Log;
 
 import com.uni.stuttgart.ipvs.androidgateway.gateway.GatewayService;
 import com.uni.stuttgart.ipvs.androidgateway.gateway.IGatewayService;
+import com.uni.stuttgart.ipvs.androidgateway.gateway.IScheduler;
 import com.uni.stuttgart.ipvs.androidgateway.gateway.PBluetoothGatt;
 import com.uni.stuttgart.ipvs.androidgateway.helper.PowerEstimator;
-import com.uni.stuttgart.ipvs.androidgateway.gateway.mcdm.ANP;
+import com.uni.stuttgart.ipvs.androidgateway.gateway.mcdm.WSM;
 import com.uni.stuttgart.ipvs.androidgateway.thread.ExecutionTask;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
 
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 
-public class PriorityBasedWithANP {
+public class PriorityBasedWithWSM implements IScheduler {
 
-    private static final int SCAN_TIME = 10000; // set scanning and reading time to 10 seoonds
-    private static final int PROCESSING_TIME = 60000; // set processing time to 60 seconds
-    private static final int NUMBER_OF_MAX_CONNECT_DEVICES = 10; // set max 10 devices connect before listening to disconnection time
+    private int SCAN_TIME; // set scanning and reading time to 10 seoonds
+    private int SCAN_TIME_HALF; // set scanning and reading time half of original scan time
+    private int PROCESSING_TIME; // set processing time to 60 seconds
+    private int TIME_MEASURE_POWER = 500; // measure power every 0.5 seconds
 
     private ScheduledThreadPoolExecutor scheduler;
     private ScheduledThreadPoolExecutor scheduler2;
+    private ScheduledThreadPoolExecutor schedulerPowerMeasure;
     private ScheduledFuture<?> future;
     private ScheduledFuture<?> future2;
-    private Thread threadMeasurePower;
 
-    private Thread sleepThread;
     private IGatewayService iGatewayService;
     private PowerEstimator powerEstimator;
 
     private boolean mScanning;
     private Context context;
     private boolean mProcessing;
-    private boolean isBroadcastRegistered;
+    private boolean mConnecting;
 
     private int cycleCounter = 0;
     private int maxConnectTime = 0;
     private long powerUsage = 0;
-    private int connectCounter = 0;
 
     private ExecutionTask<Void> executionTask;
 
-    public PriorityBasedWithANP(Context context, boolean mProcessing, IGatewayService iGatewayService, ExecutionTask<Void> executionTask) {
+    public PriorityBasedWithWSM(Context context, boolean mProcessing, IGatewayService iGatewayService, ExecutionTask<Void> executionTask) {
         this.context = context;
         this.mProcessing = mProcessing;
         this.iGatewayService = iGatewayService;
         this.executionTask = executionTask;
+
+        try {
+            this.SCAN_TIME = iGatewayService.getTimeSettings("ScanningTime");
+            this.SCAN_TIME_HALF = iGatewayService.getTimeSettings("ScanningTime2");
+            this.PROCESSING_TIME = iGatewayService.getTimeSettings("ProcessingTime");
+        } catch (RemoteException e) {
+            e.printStackTrace();
+        }
     }
 
     public void stop() {
+        mConnecting = false;
         mProcessing = false;
-        future.cancel(true);future2.cancel(true);
-        scheduler.shutdownNow();scheduler2.shutdownNow();
+        future.cancel(true);
+        future2.cancel(true);
+        scheduler.shutdownNow();
+        scheduler2.shutdownNow();
     }
 
     public void start() {
         try {
-            isBroadcastRegistered = false;
+            mConnecting = false;
             powerEstimator = new PowerEstimator(context);
 
             scheduler = executionTask.scheduleWithThreadPoolExecutor(new FPStartScanning(), 0, PROCESSING_TIME + 1, MILLISECONDS);
@@ -83,56 +97,71 @@ public class PriorityBasedWithANP {
             try {
                 cycleCounter++;
                 iGatewayService.setCycleCounter(cycleCounter);
-                if(cycleCounter > 1) {broadcastClrScrn();}
+                if (cycleCounter > 1) {
+                    broadcastClrScrn();
+                }
                 broadcastUpdate("Start new cycle");
-
                 broadcastUpdate("Cycle number " + cycleCounter);
-                mProcessing = true;
+
                 boolean isDataExist = iGatewayService.checkDevice(null);
                 if (isDataExist) {
                     // Devices are listed in DB
                     List<String> devices = iGatewayService.getListActiveDevices();
-                    for (String device : devices) { iGatewayService.startScanKnownDevices(device);/*iGatewayService.addQueueScanning(device, null, 0, BluetoothLeDevice.FIND_LE_DEVICE, null, 0);*/ }
+                    // search for known device listed in database
+                    for (String device : devices) {
+                        iGatewayService.startScanKnownDevices(device);
+                    }
 
                     // do normal scanning only for half of normal scanning time
-
-                    iGatewayService.startScan(SCAN_TIME / 2);
+                    iGatewayService.startScan(SCAN_TIME_HALF);
                     iGatewayService.stopScanning();
                     mScanning = iGatewayService.getScanState();
+
+                    if (!mProcessing) {
+                        future.cancel(true);
+                        Thread.currentThread().interrupt();
+                        return;
+                    }
 
                     waitThread(100);
 
                     if (!mProcessing) {
-                        future.cancel(false);
+                        future.cancel(true);
+                        Thread.currentThread().interrupt();
                         return;
                     }
                     connectFP();
                 } else {
                     // do normal scanning
-
                     iGatewayService.startScan(SCAN_TIME);
                     iGatewayService.stopScanning();
-
-                    //iGatewayService.execScanningQueue();
                     mScanning = iGatewayService.getScanState();
 
                     if (!mProcessing) {
-                        future.cancel(false);
+                        future.cancel(true);
+                        Thread.currentThread().interrupt();
                         return;
                     }
 
                     waitThread(100);
 
                     if (!mProcessing) {
-                        future.cancel(false);
+                        future.cancel(true);
+                        Thread.currentThread().interrupt();
                         return;
                     }
                     connectRR();
                 }
 
-
+                if (!mProcessing) {
+                    future.cancel(true);
+                    Thread.currentThread().interrupt();
+                    return;
+                }
             } catch (Exception e) {
                 e.printStackTrace();
+            } finally {
+                Log.d("Thread", "Thread FEPWSM " + Thread.currentThread().getId() + " is interrupted");
             }
         }
 
@@ -152,23 +181,30 @@ public class PriorityBasedWithANP {
                 }
 
                 // do connecting by Round Robin
-                for (final BluetoothDevice device : scanResults) {
-                    broadcastServiceInterface("Start service interface");
+                for (BluetoothDevice device : new ArrayList<BluetoothDevice>(scanResults)) {
+                    mConnecting = true;
                     iGatewayService.updateDatabaseDeviceState(device, "inactive");
+                    broadcastServiceInterface("Start service interface");
 
-                    setMeasurePower("Start", device);
+                    startStopPowerMeasure(device, "Start");
+
                     PBluetoothGatt parcelBluetoothGatt = iGatewayService.doConnecting(device.getAddress());
+                    schedulerPowerMeasure = executionTask.scheduleWithThreadPoolExecutor(doMeasurePower(), 0, TIME_MEASURE_POWER, TimeUnit.MILLISECONDS);
 
                     // set timer to xx seconds
                     waitThread(maxConnectTime);
-                    if (!mProcessing) { return; }
 
                     broadcastUpdate("Wait time finished, disconnected...");
                     iGatewayService.doDisconnected(parcelBluetoothGatt, "GatewayController");
 
-                    setMeasurePower("Stop", device);
+                    schedulerPowerMeasure.shutdownNow();
+                    startStopPowerMeasure(device, "Stop");
 
-                    iGatewayService.updateDatabaseDevicePowerUsage(device.getAddress(), powerUsage);
+                    if (!mProcessing) {
+                        future.cancel(true);
+                        Thread.currentThread().interrupt();
+                        return;
+                    }
                 }
             } catch (Exception e) {
                 e.printStackTrace();
@@ -178,15 +214,14 @@ public class PriorityBasedWithANP {
         // 2nd and so on iterations, connect using Fixed Priority Scheduling by device ranking
         private void connectFP() {
             try {
-                /*registerBroadcast(); // start listening to disconnected Gatt and or finished read data*/
                 List<BluetoothDevice> scanResults = iGatewayService.getScanResults();
                 iGatewayService.setScanResultNonVolatile(scanResults);
 
                 Map<BluetoothDevice, Double> mapRankedDevices;
                 if (scanResults.size() != 0) {
                     broadcastUpdate("\n");
-                    broadcastUpdate("Start ranking device with ANP algorithm...");
-                    mapRankedDevices = doRankDeviceANP(scanResults);
+                    broadcastUpdate("Start ranking device with WSM algorithm...");
+                    mapRankedDevices = doRankDeviceWSM(scanResults);
                     broadcastUpdate("Finish ranking device...");
                 } else {
                     broadcastUpdate("No nearby device(s) available...");
@@ -196,14 +231,15 @@ public class PriorityBasedWithANP {
                 // calculate timer for connection (to obtain Round Robin Scheduling)
                 int remainingTime = PROCESSING_TIME - SCAN_TIME;
 
-                if(mapRankedDevices.size() > 0) {
+                if (mapRankedDevices.size() > 0) {
                     broadcastUpdate("\n");
                     maxConnectTime = remainingTime / mapRankedDevices.size();
                     broadcastUpdate("Connecting to " + mapRankedDevices.size() + " device(s)");
                     broadcastUpdate("Maximum connection time for all devices is " + maxConnectTime / 1000 + " s");
                     broadcastUpdate("\n");
 
-                    connect(mapRankedDevices, remainingTime);
+                    connect(mapRankedDevices);
+
                 } else {
                     broadcastUpdate("No nearby device(s) available");
                     return;
@@ -216,26 +252,31 @@ public class PriorityBasedWithANP {
         }
 
         // implementation of connect after ranking the devices
-        private void connect(Map<BluetoothDevice, Double> mapRankedDevices, int remainingTime) {
+        private void connect(Map<BluetoothDevice, Double> mapRankedDevices) {
             try {
                 for (Map.Entry entry : mapRankedDevices.entrySet()) {
+                    mConnecting = true;
                     BluetoothDevice device = (BluetoothDevice) entry.getKey();
                     iGatewayService.updateDatabaseDeviceState(device, "inactive");
 
-                    setMeasurePower("Start", device);
-                    Thread connectingThread = executionTask.executeRunnableInThread(doConnecting(device.getAddress()), "Thread Connecting " + device.getAddress(), Thread.MAX_PRIORITY);
+                    startStopPowerMeasure(device, "Start");
+
+                    PBluetoothGatt parcelBluetoothGatt = iGatewayService.doConnecting(device.getAddress());
+                    schedulerPowerMeasure = executionTask.scheduleWithThreadPoolExecutor(doMeasurePower(), 0, TIME_MEASURE_POWER, TimeUnit.MILLISECONDS);
 
                     // set timer to xx seconds
                     waitThread(maxConnectTime);
-                    if (!mProcessing) { return; }
+                    if (!mProcessing) {
+                        future.cancel(true);
+                        Thread.currentThread().interrupt();
+                        return;
+                    }
 
                     broadcastUpdate("Wait time finished, disconnected...");
-                    iGatewayService.doDisconnected(iGatewayService.getCurrentGatt(), "GatewayController");
-                    waitThread(10);
-                    executionTask.interruptThread(connectingThread);
+                    iGatewayService.doDisconnected(parcelBluetoothGatt, "GatewayController");
 
-                    setMeasurePower("Stop", device);
-                    iGatewayService.updateDatabaseDevicePowerUsage(device.getAddress(), powerUsage);
+                    schedulerPowerMeasure.shutdownNow();
+                    startStopPowerMeasure(device, "Stop");
                 }
             } catch (Exception e) {
                 e.printStackTrace();
@@ -243,13 +284,13 @@ public class PriorityBasedWithANP {
 
         }
 
-        // implementation of Ranking Devices based on ANP
-        private Map<BluetoothDevice, Double> doRankDeviceANP(List<BluetoothDevice> devices) {
+        // implementation of Ranking Devices based on WSM
+        private Map<BluetoothDevice, Double> doRankDeviceWSM(List<BluetoothDevice> devices) {
             Map<BluetoothDevice, Double> result = new ConcurrentHashMap<>();
             try {
-                ANP anp = new ANP(devices, iGatewayService, powerEstimator.getBatteryRemainingPercent());
+                WSM wsm = new WSM(devices, iGatewayService, powerEstimator.getBatteryRemainingPercent());
                 broadcastUpdate("Sorting devices by their priorities...");
-                result = anp.call();
+                result = wsm.call();
             } catch (Exception e) {
                 e.printStackTrace();
             }
@@ -262,6 +303,7 @@ public class PriorityBasedWithANP {
         public void run() {
             if (!mProcessing) {
                 future2.cancel(true);
+                Thread.currentThread().interrupt();
                 return;
             }
             broadcastUpdate("Update all device states...");
@@ -272,7 +314,7 @@ public class PriorityBasedWithANP {
                     e.printStackTrace();
                 }
             } else {
-                future2.cancel(false);
+                future2.cancel(true);
             }
         }
     }
@@ -290,8 +332,14 @@ public class PriorityBasedWithANP {
             public void run() {
                 try {
                     long currentNow = powerEstimator.getCurrentNow();
-                    if (currentNow < 0) { currentNow = currentNow * -1; }
+                    if (currentNow < 0) {
+                        currentNow = currentNow * -1;
+                    }
                     powerUsage = powerUsage + (currentNow * new Long(powerEstimator.getVoltageNow()));
+                    if (!mConnecting) {
+                        schedulerPowerMeasure.shutdownNow();
+                        return;
+                    }
                 } catch (Exception e) {
                     e.printStackTrace();
                 }
@@ -299,36 +347,27 @@ public class PriorityBasedWithANP {
         };
     }
 
-    private synchronized void setMeasurePower(String type, BluetoothDevice device) {
+    private synchronized void startStopPowerMeasure(BluetoothDevice mDevice, String type) {
         switch (type) {
             case "Start":
                 powerUsage = 0;
                 powerEstimator.start();
-                threadMeasurePower = executionTask.executeRunnableInThread(doMeasurePower(), "Thread Power Measure " + device.getAddress(), Thread.MIN_PRIORITY);
                 break;
             case "Stop":
-                threadMeasurePower.interrupt();
+                mConnecting = false;
                 powerEstimator.stop();
+
+                try {
+                    iGatewayService.updateDatabaseDevicePowerUsage(mDevice.getAddress(), powerUsage);
+                } catch (RemoteException e) {
+                    e.printStackTrace();
+                }
                 break;
         }
     }
 
-    private Runnable doConnecting(final String macAddress) {
-        return new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    iGatewayService.doConnect(macAddress);
-                } catch (RemoteException e) {
-                    e.printStackTrace();
-                }
-            }
-        };
-    }
-
     private void waitThread(long time) {
         try {
-            sleepThread = Thread.currentThread();
             Thread.sleep(time);
         } catch (InterruptedException e) {
             e.printStackTrace();
@@ -357,5 +396,4 @@ public class PriorityBasedWithANP {
             context.sendBroadcast(intent);
         }
     }
-
 }
