+++ app/src/main/java/com/uni/stuttgart/ipvs/androidgateway/gateway/scheduler/Semaphore.java
@@ -1,38 +1,41 @@
-package com.uni.stuttgart.ipvs.androidgateway.gateway.scheduling;
+package com.uni.stuttgart.ipvs.androidgateway.gateway.scheduler;
 
 import android.bluetooth.BluetoothDevice;
 import android.content.Context;
 import android.content.Intent;
 import android.os.RemoteException;
-import android.util.Log;
 
 import com.uni.stuttgart.ipvs.androidgateway.gateway.GatewayService;
 import com.uni.stuttgart.ipvs.androidgateway.gateway.IGatewayService;
+import com.uni.stuttgart.ipvs.androidgateway.gateway.IScheduler;
+import com.uni.stuttgart.ipvs.androidgateway.helper.PowerEstimator;
+import com.uni.stuttgart.ipvs.androidgateway.thread.EExecutionType;
 import com.uni.stuttgart.ipvs.androidgateway.thread.ExecutionTask;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.Future;
 
-// implementation of Scheduling using Exhaustive Polling
-public class ExhaustivePolling {
+// implementation of Semaphore Scheduling Gateway Controller
 
-    private IGatewayService iGatewayService;
-
+public class Semaphore implements IScheduler {
     private int SCAN_TIME; // set scanning and reading time to 10 seoonds
-    private int SCAN_TIME_OTHER; // set scanning and reading time half of original scan time
+    private int SCAN_TIME_HALF;
     private int PROCESSING_TIME; // set processing time to 60 seconds
 
+    private IGatewayService iGatewayService;
+    private PowerEstimator mServicePE;
+
+    private boolean mBoundPE;
     private boolean mScanning;
     private Context context;
     private boolean mProcessing;
-
     private int cycleCounter = 0;
+
     private ExecutionTask<Void> executionTask;
-    private Future<?> futureEP;
+    private Future<Void> future;
 
-
-    public ExhaustivePolling(Context context, boolean mProcessing, IGatewayService iGatewayService, ExecutionTask<Void> executionTask) {
+    public Semaphore(Context context, boolean mProcessing, IGatewayService iGatewayService, ExecutionTask<Void> executionTask) {
         this.context = context;
         this.mProcessing = mProcessing;
         this.iGatewayService = iGatewayService;
@@ -40,89 +43,72 @@ public class ExhaustivePolling {
 
         try {
             this.SCAN_TIME = iGatewayService.getTimeSettings("ScanningTime");
-            this.SCAN_TIME_OTHER = iGatewayService.getTimeSettings("ScanningTime2");
+            this.SCAN_TIME_HALF = iGatewayService.getTimeSettings("ScanningTime2");
             this.PROCESSING_TIME = iGatewayService.getTimeSettings("ProcessingTime");
         } catch (RemoteException e) {
             e.printStackTrace();
         }
     }
-
+    
     public void stop() {
         mProcessing = false;
-        futureEP.cancel(true);
+        future.cancel(true);
     }
 
     public void start() {
-        mProcessing = true;
-        futureEP = executionTask.submitRunnable(new RunEPScheduling());
+        executionTask.setExecutionType(EExecutionType.SINGLE_THREAD_POOL);
+        future = executionTask.submitRunnable(new SemaphoreSchedulling());
     }
 
-    private class RunEPScheduling implements Runnable {
+    private class SemaphoreSchedulling implements Runnable {
+
         @Override
         public void run() {
             while (!Thread.currentThread().isInterrupted() && mProcessing) {
-                // do Exhaustive Polling Part
                 cycleCounter++;
-                if (cycleCounter > 1) {
-                    broadcastClrScrn();
-                }
+                if(cycleCounter > 1) {broadcastClrScrn();}
                 broadcastUpdate("\n");
-                broadcastUpdate("Start new cycle");
+                broadcastUpdate("Start new cycle...");
                 broadcastUpdate("Cycle number " + cycleCounter);
 
                 try {
                     iGatewayService.setCycleCounter(cycleCounter);
-                    boolean isDataExist = iGatewayService.checkDevice(null);
-                    if (isDataExist) {
-                        // Devices are listed in DB
-                        List<String> devices = iGatewayService.getListActiveDevices();
-                        // search for known device listed in database
-                        for (String device : devices) {
-                            iGatewayService.startScanKnownDevices(device);
-                        }
 
-                        // do normal scanning only for half of normal scanning time
-                        iGatewayService.startScan(SCAN_TIME_OTHER);
-                        iGatewayService.stopScan();
-                        mScanning = iGatewayService.getScanState();
-                    } else {
-                        // do normal scanning
+                    iGatewayService.startScan(SCAN_TIME);
+                    iGatewayService.stopScanning();
 
-                        iGatewayService.startScan(SCAN_TIME);
-                        iGatewayService.stopScan();
-                        mScanning = iGatewayService.getScanState();
+                    mScanning = iGatewayService.getScanState();
+
+                    if (!mProcessing) {
+                        future.cancel(true);
+                        Thread.currentThread().interrupt();
+                        return;
                     }
 
                     List<BluetoothDevice> scanResults = iGatewayService.getScanResults();
                     iGatewayService.setScanResultNonVolatile(scanResults);
 
-                    if (!mProcessing) {
-                        futureEP.cancel(true);
-                        Thread.currentThread().interrupt();
-                        return;
-                    }
-
-                    // do Connecting by using Semaphore
+                    // do Semaphore for Connecting method
                     for (BluetoothDevice device : new ArrayList<BluetoothDevice>(scanResults)) {
+                        // only known manufacturer that will be used to connect
                         iGatewayService.broadcastServiceInterface("Start service interface");
                         iGatewayService.doConnect(device.getAddress());
 
                         if (!mProcessing) {
-                            futureEP.cancel(true);
+                            future.cancel(true);
                             Thread.currentThread().interrupt();
                             return;
                         }
+
                     }
 
                     if (!mProcessing) {
-                        futureEP.cancel(true);
+                        future.cancel(true);
                         Thread.currentThread().interrupt();
                         return;
                     }
-                } catch (Exception e) {
+                } catch (RemoteException e) {
                     e.printStackTrace();
-                } finally {
-                    Log.d("Thread", "Thread EP " + Thread.currentThread().getId() + " is interrupted");
                 }
             }
         }
